test(ModelService) :  Tests for Move Validation
TestClass for MoveValidation
