+++ app/src/test/java/com/example/stratego_app/pieces/ModelServiceTest.java
@@ -19,30 +19,7 @@ import com.example.stratego_app.model.pieces.*;
 class ModelServiceTest {
 
     private ModelService modelService = new ModelService();
-    @Test
-    public void testMovePiece_ValidMove() {
-        // Arrange: Assuming you have a valid setup
-        int startX = 2, startY = 1, endX = 2, endY = 1;
-        modelService.getBoard().setField(startY, startX, new Piece(Rank.MINER));
-        boolean expected = true;
 
-        boolean result = modelService.movePiece(startX, startY, endX, endY);
-
-        // Assert: Validate that the move was successful
-        assertEquals("Valid move should be successful", expected, result);
-    }
-    @Test
-    public void testMovePiece_NotValidMove() {
-        // Arrange: Assuming you have a valid setup
-        int startX = 1, startY = 1, endX = 2, endY = 2;
-        modelService.getBoard().setField(startY, startX, new Piece(Rank.MINER));
-        boolean expected = false;
-
-        boolean result = modelService.movePiece(startX, startY, endX, endY);
-
-        // Assert: Validate that the move was successful
-        assertEquals("Move should not be successful", expected, result);
-    }
     @Test
     public void getPieceTest() {
         assertNull(modelService.getPieceAtPosition(1,1));
